Support ingress mode in the chart.
If you set ingress.enabled to True, the GUI always uses a LoadBalancer, the "face" deployment is always an "ingress", and there's a new workload called "cell" that does the actual face-workload functionality. It's done this way so that the GUI doesn't change, and so that the request into the cluster is always made with a path of "/face/".

Signed-off-by: Flynn <flynn@buoyant.io>